+++ src/main/java/gov/usgs/owi/nldi/springinit/SpringConfig.java
@@ -3,6 +3,7 @@ package gov.usgs.owi.nldi.springinit;
 import gov.usgs.owi.nldi.controllers.BaseController;
 import gov.usgs.owi.nldi.converters.ComidMessageConverter;
 import gov.usgs.owi.nldi.converters.ErrorMessageConverter;
+import gov.usgs.owi.nldi.converters.FeatureListMessageConverter;
 import gov.usgs.owi.nldi.converters.FeatureMessageConverter;
 import java.util.List;
 import org.springframework.boot.web.server.WebServerFactoryCustomizer;
@@ -55,6 +56,8 @@ public class SpringConfig implements WebMvcConfigurer {
     MediaType jsonLd = MediaType.valueOf(BaseController.MIME_TYPE_JSONLD);
     // we want our converters to take priority, so we add them to the top of the list
     messageConverters.add(0, new ComidMessageConverter(geoJson, MediaType.APPLICATION_JSON));
+    messageConverters.add(
+        0, new FeatureListMessageConverter(geoJson, jsonLd, MediaType.APPLICATION_JSON));
     messageConverters.add(
         0, new FeatureMessageConverter(geoJson, jsonLd, MediaType.APPLICATION_JSON));
     messageConverters.add(0, new ErrorMessageConverter(MediaType.ALL));