+++ src/test/java/gov/usgs/owi/nldi/controllers/LinkedDataControllerDataSourceIT.java
@@ -41,6 +41,18 @@ public class LinkedDataControllerDataSourceIT extends BaseControllerIT {
   }
 
   @Test
+  public void getWqpJsonLdTest() throws Exception {
+    assertEntity(
+        restTemplate,
+        "/linked-data/wqp/USGS-05427880/navigation/UM/wqp?distance=10",
+        HttpStatus.OK.value(),
+        BaseController.MIME_TYPE_JSONLD,
+        getCompareFile(RESULT_FOLDER, "getWqpJsonLdTest.json"),
+        true,
+        false);
+  }
+
+  @Test
   public void getWqpUtTestDistance() throws Exception {
     assertEntity(
         restTemplate,
@@ -48,18 +60,6 @@ public class LinkedDataControllerDataSourceIT extends BaseControllerIT {
         HttpStatus.OK.value(),
         BaseController.MIME_TYPE_GEOJSON,
         getCompareFile(RESULT_FOLDER, "getWqpUtTestDistance.json"),
-        true,
-        false);
-  }
-
-  @Test
-  public void getWqpUtTestDistanceEmpty() throws Exception {
-    assertEntity(
-        restTemplate,
-        "/linked-data/wqp/USGS-05427880/navigation/UT/wqp?distance=",
-        HttpStatus.OK.value(),
-        BaseController.MIME_TYPE_GEOJSON,
-        getCompareFile(RESULT_FOLDER, "getWqpUtTestDistanceEmpty.json"),
         true,
         false);
   }